Upload process for .fit file and subsequent route saving completed

+++ HikingTrailService.Application/DTOs/FitFileDataEntityDto.cs
@@ -4,7 +4,7 @@ public record FitFileDataEntityDto
 {
     public required Guid HikingTrailCode { get; set; }
     public required string Name { get; set; }
-    public required double Distance { get; set; }
+    public required int Distance { get; set; }
     public required DateTime StartTime { get; set; }
     public required DateTime EndTime { get; set; }
     public required double Duration { get; set; }
